+++ persimmon/website/views.py
@@ -1,20 +1,15 @@
 from decimal import Decimal
 from datetime import datetime
-import json
-import pydantic
 
 from django.db import transaction
-from django.http import HttpResponse, HttpRequest, Http404, HttpResponseBadRequest, JsonResponse, HttpResponseForbidden
+from django.http import HttpResponse, Http404
 from django.contrib.auth import authenticate, login as django_login, logout as django_logout
 from django import forms, urls
 from django.template.response import TemplateResponse
 
 from .models import User, AccountType, BankAccount, EmployeeLevel, ApprovalStatus, BankStatements
 from .common import make_user
-from .middleware import HalfCsrfViewMiddleware
-
-
-MAX_REQUEST_LENGTH = 4096
+from .middleware import api_function
 
 
 def current_user(request, required_auth=EmployeeLevel.CUSTOMER, expect_not_logged_in=False):
@@ -31,76 +26,6 @@ def current_user(request, required_auth=EmployeeLevel.CUSTOMER, expect_not_logge
         raise Http404("API unavailable to current authentication")
     return user
 
-
-def api_function(func):
-    """
-    This is a magic function which will automatically deserialize POST request data from JSON and serialize the response
-    back to JSON. Additionally, it will typecheck the JSON object structure against the function's type annotations.
-    Use it by @annotating your view functions with it.
-    """
-    gen_namespace = {name: NotImplemented for name in func.__annotations__}
-    gen_namespace['__annotations__'] = func.__annotations__
-    gen_namespace['Config'] = ApiGenConfig
-    request_type = type(func.__name__ + '_args', (pydantic.BaseModel,), gen_namespace)
-
-    def inner(request: HttpRequest, *args, **kwargs):
-        if request.method != 'POST':
-            return HttpResponseBadRequest("Must be a POST request")
-
-        request_data = request.read(MAX_REQUEST_LENGTH + 1)
-        if len(request_data) == MAX_REQUEST_LENGTH + 1:
-            return HttpResponseBadRequest("Request too large")
-        if len(request_data) == 0:
-            request_data = b''
-
-        try:
-            request_data_dict = json.loads(request_data)
-        except (json.JSONDecodeError, UnicodeDecodeError):
-            return HttpResponseBadRequest("Data encoding error")
-
-        token = request_data_dict.pop('csrfmiddlewaretoken', None)
-        if token is None:
-            return HttpResponseForbidden("No CSRF token provided")
-        if not HalfCsrfViewMiddleware.check_token(request, token):
-            return HttpResponseForbidden("CSRF protection triggered")
-
-        if not isinstance(request_data_dict, dict):
-            # force an error in the next stanza
-            request_data_dict = {'_': None}
-
-        try:
-            request_data_validated = request_type(**request_data_dict)
-        except pydantic.ValidationError:
-            return HttpResponseBadRequest("Data form error: expecting " + str(func.__annotations__))
-
-        kwargs.update(request_data_validated.dict())
-        result = func(request, *args, **kwargs)
-        if result is None:
-            result = {}
-        if isinstance(result, HttpResponse):
-            return result
-        return JsonResponse(result, safe=False, json_dumps_params=dict(default=encode_extra))
-
-    return inner
-
-
-class ApiGenConfig:
-    extra = pydantic.Extra.forbid
-    validate_all = True
-
-
-def encode_extra(thing):
-    if isinstance(thing, Decimal):
-        return str(thing)
-    if isinstance(thing, datetime):
-        return thing.isoformat()
-
-    raise TypeError(f"Object of type {thing.__class__.__name__} is not serializable")
-
-
-######
-## actual views begin here
-######
 
 # checks if user with username has same employeelevel as level
 def security_check(request, myusername, level):
@@ -374,8 +299,8 @@ def reset_password(request, email: str):
     current_user(request, expect_not_logged_in=True)
     try:
         User.objects.get(django_user__email=email)
-    except User.DoesNotExist as e:
-        raise Http404("No such email in our databases...") from e
+    except User.DoesNotExist as exc:
+        raise Http404("No such email in our databases...") from exc
     # TODO send an email here... lol
     return {}
 
@@ -395,5 +320,3 @@ def reset_password_sent(request):
     current_user(request, expect_not_logged_in=True)
 
     return TemplateResponse(request, 'pages/reset_password_sent.html', {})
-
-
