+++ pylintrc
@@ -145,6 +145,8 @@ disable=print-statement,
         too-few-public-methods,
         too-many-statements,
         too-many-arguments,
+        too-many-return-statements,
+        fixme
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
